# web/player/openapi.json
{"openapi":"3.1.0","info":{"title":"apiservice","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"summary":"Health Check","description":"Basic health check endpoint","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health/detailed":{"get":{"summary":"Detailed Health Check","description":"Detailed health check endpoint","operationId":"detailed_health_check_api_v1_health_detailed_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/web-pages":{"get":{"summary":"Read Web Pages","operationId":"read_web_pages_api_v1_web_pages_get","parameters":[{"name":"url","in":"query","required":true,"schema":{"type":"string","title":"Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/FAWebPage"},{"type":"null"}],"title":"Response Read Web Pages Api V1 Web Pages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scraper-run":{"post":{"summary":"Scraper Run","operationId":"scraper_run_api_v1_scraper_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScraperRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FAScraperStats"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/summarizer-run":{"post":{"summary":"Summarizer Run","operationId":"summarizer_run_api_v1_summarizer_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizerRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FAScraperStats"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scraper-stop":{"post":{"summary":"Scraper Stop","operationId":"scraper_stop_api_v1_scraper_stop_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/web-page-seeds":{"get":{"summary":"Read Web Page Seeds","operationId":"read_web_page_seeds_api_v1_web_page_seeds_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FAWebPageSeed"},"type":"array","title":"Response Read Web Page Seeds Api V1 Web Page Seeds Get"}}}}}},"post":{"summary":"Upsert Web Page Seed","operationId":"upsert_web_page_seed_api_v1_web_page_seeds_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertWebPageSeedRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/frontend-audio-results-similar-for-text":{"post":{"summary":"Frontend Audios Similar For Text","operationId":"frontend_audios_similar_for_text_api_v1_frontend_audio_results_similar_for_text_post","parameters":[{"name":"text","in":"query","required":true,"schema":{"type":"string","title":"Text"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FAFrontendAudioSearchResult"},"title":"Response Frontend Audios Similar For Text Api V1 Frontend Audio Results Similar For Text Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/frontend-audios-similar-for-text":{"post":{"summary":"Frontend Audios Similar For Text","operationId":"frontend_audios_similar_for_text_api_v1_frontend_audios_similar_for_text_post","parameters":[{"name":"text","in":"query","required":true,"schema":{"type":"string","title":"Text"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FAFrontendAudio"},"title":"Response Frontend Audios Similar For Text Api V1 Frontend Audios Similar For Text Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/frontend-audio-for-url":{"post":{"summary":"Frontend Audio For Url","operationId":"frontend_audio_for_url_api_v1_frontend_audio_for_url_post","parameters":[{"name":"url","in":"query","required":true,"schema":{"type":"string","title":"Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/FAFrontendAudio"},{"type":"null"}],"title":"Response Frontend Audio For Url Api V1 Frontend Audio For Url Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/frontend-audio-play":{"post":{"summary":"Frontend Audio Play","operationId":"frontend_audio_play_api_v1_frontend_audio_play_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"audio_id","in":"query","required":true,"schema":{"type":"string","title":"Audio Id"}},{"name":"duration_seconds","in":"query","required":true,"schema":{"type":"integer","title":"Duration Seconds"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FADomainStats":{"properties":{"domain":{"type":"string","title":"Domain"},"frequent_subpaths":{"additionalProperties":{"type":"integer"},"type":"object","title":"Frequent Subpaths"}},"type":"object","required":["domain","frequent_subpaths"],"title":"FADomainStats"},"FAFrontendAudio":{"properties":{"normalized_url_hash":{"type":"string","title":"Normalized Url Hash"},"normalized_url":{"type":"string","title":"Normalized Url"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"audio_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Audio Text"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"audio_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Audio Url"},"author_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author Id"},"channel_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Channel Id"},"published_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Published At"},"uploaded_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Uploaded At"},"duration":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration"},"topics":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Topics"},"similarity_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Similarity Score"}},"type":"object","required":["normalized_url_hash","normalized_url","title","description","audio_text","image_url","audio_url","author_id","channel_id","published_at","uploaded_at","duration","topics"],"title":"FAFrontendAudio"},"FAFrontendAudioSearchResult":{"properties":{"normalized_url_hash":{"type":"string","title":"Normalized Url Hash"},"similarity_score":{"type":"number","title":"Similarity Score"}},"type":"object","required":["normalized_url_hash","similarity_score"],"title":"FAFrontendAudioSearchResult"},"FAScraperStats":{"properties":{"initiated_urls_count":{"type":"integer","title":"Initiated Urls Count"},"requested_urls_count":{"type":"integer","title":"Requested Urls Count"},"completed_urls_count":{"type":"integer","title":"Completed Urls Count"},"domain_stats":{"additionalProperties":{"$ref":"#/components/schemas/FADomainStats"},"type":"object","title":"Domain Stats"}},"type":"object","required":["initiated_urls_count","requested_urls_count","completed_urls_count","domain_stats"],"title":"FAScraperStats"},"FAWebPage":{"properties":{"normalized_url_hash":{"type":"string","title":"Normalized Url Hash"},"normalized_url":{"type":"string","title":"Normalized Url"},"url":{"type":"string","title":"Url"},"status_code":{"type":"integer","title":"Status Code"},"headers":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Headers"},"content":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"Content"},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"},"content_charset":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Charset"},"metadata_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Metadata Title"},"metadata_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Metadata Description"},"metadata_image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Metadata Image Url"},"metadata_published_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Metadata Published At"},"canonical_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Canonical Url"},"outgoing_urls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Outgoing Urls"},"visible_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Visible Text"},"sitemap_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sitemap Url"},"robots_content":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Robots Content"},"text_chunks":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Text Chunks"}},"type":"object","required":["normalized_url_hash","normalized_url","url","status_code","headers","content","content_type","content_charset","metadata_title","metadata_description","metadata_image_url","metadata_published_at","canonical_url","outgoing_urls","visible_text","sitemap_url","robots_content","text_chunks"],"title":"FAWebPage"},"FAWebPageSeed":{"properties":{"normalized_url_hash":{"type":"string","title":"Normalized Url Hash"},"normalized_url":{"type":"string","title":"Normalized Url"},"url":{"type":"string","title":"Url"},"max_depth":{"type":"integer","title":"Max Depth"},"url_patterns":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Url Patterns"},"use_headless_browser":{"type":"boolean","title":"Use Headless Browser"},"allowed_domains":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Allowed Domains"}},"type":"object","required":["normalized_url_hash","normalized_url","url","max_depth","url_patterns","use_headless_browser","allowed_domains"],"title":"FAWebPageSeed"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ScraperRunRequest":{"properties":{"url":{"type":"string","title":"Url"},"max_depth":{"type":"integer","title":"Max Depth"},"no_cache":{"type":"boolean","title":"No Cache"}},"type":"object","required":["url","max_depth","no_cache"],"title":"ScraperRunRequest"},"SummarizerRunRequest":{"properties":{"url":{"type":"string","title":"Url"}},"type":"object","required":["url"],"title":"SummarizerRunRequest"},"UpsertWebPageSeedRequest":{"properties":{"normalized_url_hash":{"type":"string","title":"Normalized Url Hash"},"normalized_url":{"type":"string","title":"Normalized Url"},"url":{"type":"string","title":"Url"},"max_depth":{"type":"integer","title":"Max Depth"},"url_patterns":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Url Patterns"},"use_headless_browser":{"type":"boolean","title":"Use Headless Browser"},"allowed_domains":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Allowed Domains"}},"type":"object","required":["normalized_url_hash","normalized_url","url","max_depth","url_patterns","use_headless_browser","allowed_domains"],"title":"UpsertWebPageSeedRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}